+++ bocadillo/sessions.py
@@ -1,0 +1,2 @@
+class MissingSecretKey(Exception):
+    """Raised when no cookie siggnature secret key was found."""